refactor: 상품상세관련 컴포넌트 폴더 및 파일  common/ 으로 이동
