+++ src/components/common/product/option/option-list/SelectedOption.types.ts
@@ -1,0 +1,10 @@
+import type { OptionItem } from '../option-dropdown/OptionItem';
+
+export interface SelectedOptionProps {
+  option: OptionItem;
+  optionNumber?: number;
+  quantity?: number;
+  onIncrease?: () => void;
+  onDecrease?: () => void;
+  onRemove?: () => void;
+}
