+++ src/components/common/product/option/option-quantity-button/OptionQuantity.types.ts
@@ -1,0 +1,6 @@
+export interface OptionQuantityProps {
+  quantity?: number;
+  onIncrease?: () => void;
+  onDecrease?: () => void;
+  disabled?: boolean;
+}
