+++ src/components/common/product/option/option-quantity-button/OptionQuantity.utils.ts
@@ -1,0 +1,12 @@
+import type { OptionQuantityProps } from './OptionQuantity.types';
+
+export const getOptionQuantityProps = (
+  props: OptionQuantityProps
+): Required<OptionQuantityProps> => {
+  return {
+    quantity: props.quantity ?? 1,
+    onIncrease: props.onIncrease ?? (() => {}),
+    onDecrease: props.onDecrease ?? (() => {}),
+    disabled: props.disabled ?? false,
+  };
+};
