--- src/components/domain/product/option/option-quantity-button/OptionQuantity.types.ts
@@ -1,6 +1,0 @@
-export interface OptionQuantityProps {
-  quantity?: number;
-  onIncrease?: () => void;
-  onDecrease?: () => void;
-  disabled?: boolean;
-}
