--- src/components/domain/product/option/option-quantity-button/OptionQuantity.utils.ts
@@ -1,12 +1,0 @@
-import type { OptionQuantityProps } from './OptionQuantity.types';
-
-export const getOptionQuantityProps = (
-  props: OptionQuantityProps
-): Required<OptionQuantityProps> => {
-  return {
-    quantity: props.quantity ?? 1,
-    onIncrease: props.onIncrease ?? (() => {}),
-    onDecrease: props.onDecrease ?? (() => {}),
-    disabled: props.disabled ?? false,
-  };
-};
